requirements_old.txt
PyYAML
altgraph
babel
cffi
chardet
charset_normalizer
contourpy
cryptography    
cycler
et_xmlfile
fonttools
iniconfig
kiwisolver
matplotlib
numpy
openpyxl
packaging
pandas
pdfminer.six
pdfplumber
Pillow
pluggy
pycparser
Pygments
pyinstaller
pyinstaller-hooks-contrib
pyparsing
PyPDF2
pypdfium2
pytest
python-dateutil
pytz
reportlab
seaborn
six
tkcalendar
ttkthemes
tzdata
validators
xlrd
gspread
google-auth